requirements.txt
numpy==1.16.3
pandas==0.24.2
Pillow==6.0.0
scikit-learn==0.19.2
scipy==1.2.1
torch==1.1.0
torchvision==0.2.2.post3
xmltodict==0.12.0
